Fix: Merge tests
